Problem: `u32` for mint and burn amount may be small for ERC20 tokens (#50)
Solution:

- Changed amount to `u64` and store it in database as `TEXT`. Fixes #49.
- Change database type from TEXT to BLOB/BYTEA and use little endian encoding to store u64

+++ solo-machine-core/src/service/ibc_service.rs
@@ -276,7 +276,7 @@ impl IbcService {
         signer: impl Signer,
         chain_id: ChainId,
         request_id: Option<String>,
-        amount: u32,
+        amount: u64,
         denom: Identifier,
         receiver: Option<String>,
         memo: String,
@@ -376,7 +376,7 @@ impl IbcService {
         signer: impl Signer,
         chain_id: ChainId,
         request_id: Option<String>,
-        amount: u32,
+        amount: u64,
         denom: Identifier,
         memo: String,
     ) -> Result<String> {
